refactor(docx-reader): split DOCX reader modules
