+++ crates/docspec-docx-reader/src/rels.rs
@@ -1,44 +1,27 @@
-use std::io::{Read, Seek};
+use std::io::Read;
 
 use docspec_core::Error;
 use quick_xml::events::Event;
 use quick_xml::XmlVersion;
-use zip::result::ZipError;
 
-pub fn find_document_path<R: Read + Seek>(
-    archive: &mut zip::ZipArchive<R>,
-) -> docspec_core::Result<String> {
-    let mut rels = archive.by_name("_rels/.rels").map_err(|err| {
-        if matches!(err, ZipError::FileNotFound) {
-            Error::Parse {
-                message: "missing _rels/.rels".to_string(),
-                position: None,
-            }
-        } else {
-            parse_error(format!("malformed ZIP: {err}"))
-        }
-    })?;
-
-    let mut bytes = Vec::new();
-    rels.read_to_end(&mut bytes).map_err(Error::from)?;
-
-    let mut reader = quick_xml::Reader::from_reader(bytes.as_slice());
+pub fn find_document_target<R: Read>(reader: R) -> docspec_core::Result<String> {
+    let mut xml_reader = quick_xml::Reader::from_reader(std::io::BufReader::new(reader));
     let mut buf = Vec::new();
     let mut element_depth: usize = 0;
 
     loop {
-        match reader.read_event_into(&mut buf) {
+        match xml_reader.read_event_into(&mut buf) {
             Ok(Event::Start(element)) => {
                 element_depth = element_depth.saturating_add(1);
                 if element.local_name().as_ref() == b"Relationship" {
-                    if let Some(target) = office_document_target(&reader, &element)? {
+                    if let Some(target) = office_document_target(&xml_reader, &element)? {
                         let document_path = target.strip_prefix('/').unwrap_or(&target).to_string();
                         return validate_document_path(&document_path);
                     }
                 }
             }
             Ok(Event::Empty(element)) if element.local_name().as_ref() == b"Relationship" => {
-                if let Some(target) = office_document_target(&reader, &element)? {
+                if let Some(target) = office_document_target(&xml_reader, &element)? {
                     let document_path = target.strip_prefix('/').unwrap_or(&target).to_string();
                     return validate_document_path(&document_path);
                 }
@@ -85,8 +68,8 @@ fn validate_document_path(document_path: &str) -> docspec_core::Result<String> {
     Ok(document_path.to_string())
 }
 
-fn office_document_target(
-    reader: &quick_xml::Reader<&[u8]>,
+fn office_document_target<R: Read>(
+    reader: &quick_xml::Reader<R>,
     element: &quick_xml::events::BytesStart<'_>,
 ) -> docspec_core::Result<Option<String>> {
     let mut rel_type = None;
@@ -125,33 +108,6 @@ fn office_document_target(
 mod tests {
     use super::*;
     use std::io::Cursor;
-    use std::io::Write as _;
-    use zip::{write::SimpleFileOptions, CompressionMethod, ZipWriter};
-
-    fn synth_zip_with_rels(rels_xml: &str) -> Result<Vec<u8>, zip::result::ZipError> {
-        let buf = Cursor::new(Vec::new());
-        let mut writer = ZipWriter::new(buf);
-        let options = SimpleFileOptions::default().compression_method(CompressionMethod::Deflated);
-        writer.start_file("_rels/.rels", options)?;
-        writer.write_all(rels_xml.as_bytes())?;
-        Ok(writer.finish()?.into_inner())
-    }
-
-    fn synth_empty_zip() -> Result<Vec<u8>, zip::result::ZipError> {
-        let buf = Cursor::new(Vec::new());
-        let writer = ZipWriter::new(buf);
-        Ok(writer.finish()?.into_inner())
-    }
-
-    fn archive_from_rels(
-        rels_xml: &str,
-    ) -> Result<zip::ZipArchive<Cursor<Vec<u8>>>, zip::result::ZipError> {
-        zip::ZipArchive::new(Cursor::new(synth_zip_with_rels(rels_xml)?))
-    }
-
-    fn archive_from_empty_zip() -> Result<zip::ZipArchive<Cursor<Vec<u8>>>, zip::result::ZipError> {
-        zip::ZipArchive::new(Cursor::new(synth_empty_zip()?))
-    }
 
     fn minimal_rels(target: &str) -> String {
         format!(
@@ -169,56 +125,22 @@ mod tests {
         }
     }
 
-    fn assert_zip_result<T, E: core::fmt::Debug>(result: Result<T, E>) -> Option<T> {
-        match result {
-            Ok(value) => Some(value),
-            Err(err) => {
-                assert_eq!(format!("{err:?}"), "expected valid ZIP");
-                None
-            }
-        }
-    }
-
     #[test]
-    fn find_document_path_returns_target_for_simple_rels() {
-        let Some(mut archive) =
-            assert_zip_result(archive_from_rels(&minimal_rels("word/document.xml")))
-        else {
-            return;
-        };
+    fn find_document_target_returns_target_for_simple_rels() {
+        let rels_xml = minimal_rels("word/document.xml");
 
-        let result = find_document_path(&mut archive);
+        let result = find_document_target(Cursor::new(rels_xml.as_bytes()));
 
         assert_document_path(result, "word/document.xml");
     }
 
     #[test]
-    fn find_document_path_errors_when_rels_missing() {
-        let Some(mut archive) = assert_zip_result(archive_from_empty_zip()) else {
-            return;
-        };
-
-        let result = find_document_path(&mut archive);
-
-        match result {
-            Err(Error::Parse { message, position }) => {
-                assert_eq!(message, "missing _rels/.rels");
-                assert_eq!(position, None);
-            }
-            other => assert_eq!(format!("{other:?}"), "expected missing rels parse error"),
-        }
-    }
-
-    #[test]
-    fn find_document_path_errors_when_no_office_document_relationship() {
+    fn find_document_target_errors_when_no_office_document_relationship() {
         let rels_xml = r#"<Relationships xmlns="http://schemas.openxmlformats.org/package/2006/relationships">
   <Relationship Id="rId1" Type="http://example.com/not-office" Target="word/document.xml"/>
 </Relationships>"#;
-        let Some(mut archive) = assert_zip_result(archive_from_rels(rels_xml)) else {
-            return;
-        };
 
-        let result = find_document_path(&mut archive);
+        let result = find_document_target(Cursor::new(rels_xml.as_bytes()));
 
         match result {
             Err(Error::Parse { message, position }) => {
@@ -233,15 +155,12 @@ mod tests {
     }
 
     #[test]
-    fn find_document_path_errors_after_balanced_nested_non_matching_rels() {
+    fn find_document_target_errors_after_balanced_nested_non_matching_rels() {
         let rels_xml = r#"<Relationships xmlns="http://schemas.openxmlformats.org/package/2006/relationships">
   <Group><Relationship Id="rId1" Type="http://example.com/not-office" Target="word/document.xml"></Relationship></Group>
 </Relationships>"#;
-        let Some(mut archive) = assert_zip_result(archive_from_rels(rels_xml)) else {
-            return;
-        };
 
-        let result = find_document_path(&mut archive);
+        let result = find_document_target(Cursor::new(rels_xml.as_bytes()));
 
         match result {
             Err(Error::Parse { message, position }) => {
@@ -256,12 +175,8 @@ mod tests {
     }
 
     #[test]
-    fn find_document_path_errors_on_unexpected_closing_element() {
-        let Some(mut archive) = assert_zip_result(archive_from_rels("</Relationships>")) else {
-            return;
-        };
-
-        let result = find_document_path(&mut archive);
+    fn find_document_target_errors_on_unexpected_closing_element() {
+        let result = find_document_target(Cursor::new("</Relationships>".as_bytes()));
 
         match result {
             Err(Error::Parse { message, position }) => {
@@ -276,12 +191,8 @@ mod tests {
     }
 
     #[test]
-    fn find_document_path_errors_on_rels_xml_parser_error() {
-        let Some(mut archive) = assert_zip_result(archive_from_rels("<Relationships><")) else {
-            return;
-        };
-
-        let result = find_document_path(&mut archive);
+    fn find_document_target_errors_on_rels_xml_parser_error() {
+        let result = find_document_target(Cursor::new("<Relationships><".as_bytes()));
 
         match result {
             Err(Error::Parse { message, position }) => {
@@ -293,58 +204,45 @@ mod tests {
     }
 
     #[test]
-    fn find_document_path_strips_leading_slash() {
-        let Some(mut archive) =
-            assert_zip_result(archive_from_rels(&minimal_rels("/word/document.xml")))
-        else {
-            return;
-        };
+    fn find_document_target_strips_leading_slash() {
+        let rels_xml = minimal_rels("/word/document.xml");
 
-        let result = find_document_path(&mut archive);
+        let result = find_document_target(Cursor::new(rels_xml.as_bytes()));
 
         assert_document_path(result, "word/document.xml");
     }
 
     #[test]
-    fn find_document_path_picks_office_document_among_multiple_relationships() {
+    fn find_document_target_picks_office_document_among_multiple_relationships() {
         let rels_xml = r#"<Relationships xmlns="http://schemas.openxmlformats.org/package/2006/relationships">
   <Relationship Id="rId1" Type="http://example.com/metadata" Target="docProps/core.xml"/>
   <Relationship Id="rId2" Type="http://schemas.openxmlformats.org/officeDocument/2006/relationships/officeDocument" Target="word/document.xml"/>
   <Relationship Id="rId3" Type="http://example.com/theme" Target="word/theme/theme1.xml"/>
 </Relationships>"#;
-        let Some(mut archive) = assert_zip_result(archive_from_rels(rels_xml)) else {
-            return;
-        };
 
-        let result = find_document_path(&mut archive);
+        let result = find_document_target(Cursor::new(rels_xml.as_bytes()));
 
         assert_document_path(result, "word/document.xml");
     }
 
     #[test]
-    fn find_document_path_accepts_non_empty_relationship_element() {
+    fn find_document_target_accepts_non_empty_relationship_element() {
         let rels_xml = r#"<Relationships xmlns="http://schemas.openxmlformats.org/package/2006/relationships">
   <Relationship Id="rId1" Type="http://schemas.openxmlformats.org/officeDocument/2006/relationships/officeDocument" Target="word/document.xml"></Relationship>
 </Relationships>"#;
-        let Some(mut archive) = assert_zip_result(archive_from_rels(rels_xml)) else {
-            return;
-        };
 
-        let result = find_document_path(&mut archive);
+        let result = find_document_target(Cursor::new(rels_xml.as_bytes()));
 
         assert_document_path(result, "word/document.xml");
     }
 
     #[test]
-    fn find_document_path_errors_on_malformed_relationship_attribute() {
+    fn find_document_target_errors_on_malformed_relationship_attribute() {
         let rels_xml = r#"<Relationships xmlns="http://schemas.openxmlformats.org/package/2006/relationships">
   <Relationship Id="rId1" Type="http://schemas.openxmlformats.org/officeDocument/2006/relationships/officeDocument" Target=word/document.xml/>
 </Relationships>"#;
-        let Some(mut archive) = assert_zip_result(archive_from_rels(rels_xml)) else {
-            return;
-        };
 
-        let result = find_document_path(&mut archive);
+        let result = find_document_target(Cursor::new(rels_xml.as_bytes()));
 
         match result {
             Err(Error::Parse { message, position }) => {
@@ -359,15 +257,12 @@ mod tests {
     }
 
     #[test]
-    fn find_document_path_errors_on_bad_attribute_entity() {
+    fn find_document_target_errors_on_bad_attribute_entity() {
         let rels_xml = r#"<Relationships xmlns="http://schemas.openxmlformats.org/package/2006/relationships">
   <Relationship Id="rId1" Type="http://schemas.openxmlformats.org/officeDocument/2006/relationships/officeDocument" Target="word/&bogus;.xml"/>
 </Relationships>"#;
-        let Some(mut archive) = assert_zip_result(archive_from_rels(rels_xml)) else {
-            return;
-        };
 
-        let result = find_document_path(&mut archive);
+        let result = find_document_target(Cursor::new(rels_xml.as_bytes()));
 
         match result {
             Err(Error::Parse { message, position }) => {
@@ -382,26 +277,19 @@ mod tests {
     }
 
     #[test]
-    fn find_document_path_tolerates_namespaced_relationship_element() {
+    fn find_document_target_tolerates_namespaced_relationship_element() {
         let rels_xml = r#"<r:Relationships xmlns:r="http://schemas.openxmlformats.org/package/2006/relationships">
   <r:Relationship Id="rId1" Type="http://schemas.openxmlformats.org/officeDocument/2006/relationships/officeDocument" Target="word/document.xml"/>
 </r:Relationships>"#;
-        let Some(mut archive) = assert_zip_result(archive_from_rels(rels_xml)) else {
-            return;
-        };
 
-        let result = find_document_path(&mut archive);
+        let result = find_document_target(Cursor::new(rels_xml.as_bytes()));
 
         assert_document_path(result, "word/document.xml");
     }
 
     #[test]
-    fn find_document_path_errors_on_malformed_rels_xml() {
-        let Some(mut archive) = assert_zip_result(archive_from_rels("<broken>")) else {
-            return;
-        };
-
-        let result = find_document_path(&mut archive);
+    fn find_document_target_errors_on_malformed_rels_xml() {
+        let result = find_document_target(Cursor::new("<broken>".as_bytes()));
 
         match result {
             Err(Error::Parse { message, position }) => {
@@ -413,14 +301,10 @@ mod tests {
     }
 
     #[test]
-    fn find_document_path_rejects_target_with_dotdot_segment() {
-        let Some(mut archive) =
-            assert_zip_result(archive_from_rels(&minimal_rels("../foo/document.xml")))
-        else {
-            return;
-        };
+    fn find_document_target_rejects_target_with_dotdot_segment() {
+        let rels_xml = minimal_rels("../foo/document.xml");
 
-        let result = find_document_path(&mut archive);
+        let result = find_document_target(Cursor::new(rels_xml.as_bytes()));
 
         match result {
             Err(Error::Parse { message, position }) => {
@@ -435,14 +319,10 @@ mod tests {
     }
 
     #[test]
-    fn find_document_path_handles_target_with_entities() {
-        let Some(mut archive) =
-            assert_zip_result(archive_from_rels(&minimal_rels("word/doc&amp;ument.xml")))
-        else {
-            return;
-        };
+    fn find_document_target_handles_target_with_entities() {
+        let rels_xml = minimal_rels("word/doc&amp;ument.xml");
 
-        let result = find_document_path(&mut archive);
+        let result = find_document_target(Cursor::new(rels_xml.as_bytes()));
         assert_document_path(result, "word/doc&ument.xml");
     }
 }
